cleaned up some endpoints
